backend_requirements.txt
Django==5.0.2
djangorestframework==3.15.2
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
python-decouple==3.8
psycopg2-binary==2.9.9
Pillow==10.2.0
celery==5.3.4
redis==5.0.1
django-filter==23.5
drf-spectacular==0.27.1 -